Move router and metrics to api package
Moving the router and metrics to separate package outside of main package
enables us to reuse it for our end-to-end tests and use the main router for
testing.
